check_latest_epoch first in each iteration regardless of prev epoch dumped; do not store progress anymore; fix the epoch_id of integration test and move the test in state_sync.rs into integration-tests
