+++ integration-tests/src/tests/client/mod.rs
@@ -9,5 +9,6 @@ mod runtimes;
 #[cfg(feature = "sandbox")]
 mod sandbox;
 mod sharding_upgrade;
+mod state_dump;
 mod undo_block;
 mod utils;
