+++ nearcore/src/state_sync.rs
@@ -197,149 +197,114 @@ async fn state_sync_dump(
     // Stop if the node is stopped.
     // Note that without this check the state dumping thread is unstoppable, i.e. non-interruptable.
     while keep_running.load(std::sync::atomic::Ordering::Relaxed) {
-        // TODO (ND-437): Start every iteration of the state dumping loop with checking if a new epoch is available.
-        let progress = chain.store().get_state_sync_dump_progress(shard_id);
-        tracing::debug!(target: "state_sync_dump", shard_id, ?progress, "Running StateSyncDump loop iteration");
-        // The `match` returns the next state of the state machine.
-        let next_state: Result<Option<StateSyncDumpProgress>, Error> = match progress {
-            Ok(Some(StateSyncDumpProgress::AllDumped { epoch_id })) => {
-                // The latest epoch was dumped. Check if a newer epoch is available.
-                check_new_epoch(
-                    Some(epoch_id),
-                    shard_id,
-                    &chain,
-                    epoch_manager.as_ref(),
-                    &shard_tracker,
-                    &account_id,
-                )
-            }
-            Err(Error::DBNotFoundErr(_)) | Ok(None) => {
-                // First invocation of this state-machine. See if at least one epoch is available for dumping.
-                check_new_epoch(
-                    None,
-                    shard_id,
-                    &chain,
-                    epoch_manager.as_ref(),
-                    &shard_tracker,
-                    &account_id,
-                )
+        tracing::debug!(target: "state_sync_dump", shard_id, "Running StateSyncDump loop iteration");
+        let latest_epoch_state = check_latest_epoch(
+            shard_id,
+            &chain,
+            epoch_manager.as_ref(),
+            &shard_tracker,
+            &account_id,
+        );
+        let next_state: Result<StateSyncDumpProgress, Error> = match latest_epoch_state {
+            Ok(StateSyncDumpProgress::AllDumped { epoch_id: _, epoch_height: _ }) => {
+                // shard is not tracked for the latest available epoch
+                continue;
             }
             Err(err) => {
-                // Something went wrong, let's retry.
-                tracing::warn!(target: "state_sync_dump", shard_id, ?err, "Failed to read the progress, will now delete and retry");
-                if let Err(err) = chain.store().set_state_sync_dump_progress(shard_id, None) {
-                    tracing::warn!(target: "state_sync_dump", shard_id, ?err, "and failed to delete the progress. Will later retry.");
-                }
-                Ok(None)
+                tracing::debug!(target: "state_sync_dump", shard_id, ?err, "check_latest_epoch failed, will retry.");
+                continue;
             }
-            Ok(Some(StateSyncDumpProgress::InProgress { epoch_id, epoch_height, sync_hash })) => {
+            Ok(StateSyncDumpProgress::InProgress { epoch_id, epoch_height, sync_hash }) => {
                 let in_progress_data = get_in_progress_data(shard_id, sync_hash, &chain);
                 match in_progress_data {
                     Err(error) => Err(error),
                     Ok((state_root, num_parts, sync_prev_hash)) => {
-                        let new_epoch_state = check_new_epoch(
-                            Some(epoch_id.clone()),
+                        let missing_parts = get_missing_part_ids_for_epoch(
                             shard_id,
-                            &chain,
-                            epoch_manager.as_ref(),
-                            &shard_tracker,
-                            &account_id,
-                        );
-                        match new_epoch_state {
-                            Err(err) => Err(err),
-                            // there's a new epoch available, next iteration will work on it
-                            Ok(Some(state)) => Ok(Some(state)),
-                            Ok(None) => {
-                                let missing_parts = get_missing_part_ids_for_epoch(
-                                    shard_id,
-                                    &chain_id,
-                                    &epoch_id,
-                                    epoch_height,
-                                    num_parts,
-                                    &external,
-                                )
-                                .await;
+                            &chain_id,
+                            &epoch_id,
+                            epoch_height,
+                            num_parts,
+                            &external,
+                        )
+                        .await;
 
-                                match missing_parts {
-                                    Err(err) => {
-                                        tracing::debug!(target: "state_sync_dump", shard_id, ?err, "get_missing_state_parts_for_epoch error");
-                                        Err(Error::Other(format!(
-                                            "get_missing_state_parts_for_epoch failed"
-                                        )))
-                                    }
-                                    Ok(parts_not_dumped) if parts_not_dumped.is_empty() => {
-                                        Ok(Some(StateSyncDumpProgress::AllDumped { epoch_id }))
-                                    }
-                                    Ok(parts_not_dumped) => {
-                                        let mut parts_to_dump = parts_not_dumped.clone();
-                                        let timer = Instant::now();
-                                        // Stop if the node is stopped.
-                                        // Note that without this check the state dumping thread is unstoppable, i.e. non-interruptable.
-                                        while keep_running
-                                            .load(std::sync::atomic::Ordering::Relaxed)
-                                            && timer.elapsed().as_secs()
-                                                <= STATE_DUMP_ITERATION_TIME_LIMIT_SECS
-                                            && !parts_to_dump.is_empty()
-                                        {
-                                            let _timer = metrics::STATE_SYNC_DUMP_ITERATION_ELAPSED
-                                                .with_label_values(&[&shard_id.to_string()])
-                                                .start_timer();
+                        match missing_parts {
+                            Err(err) => {
+                                tracing::debug!(target: "state_sync_dump", shard_id, ?err, "get_missing_state_parts_for_epoch error");
+                                Err(Error::Other(format!(
+                                    "get_missing_state_parts_for_epoch failed"
+                                )))
+                            }
+                            Ok(parts_not_dumped) if parts_not_dumped.is_empty() => {
+                                Ok(StateSyncDumpProgress::AllDumped { epoch_id, epoch_height })
+                            }
+                            Ok(parts_not_dumped) => {
+                                let mut parts_to_dump = parts_not_dumped.clone();
+                                let timer = Instant::now();
+                                // Stop if the node is stopped.
+                                // Note that without this check the state dumping thread is unstoppable, i.e. non-interruptable.
+                                while keep_running.load(std::sync::atomic::Ordering::Relaxed)
+                                    && timer.elapsed().as_secs()
+                                        <= STATE_DUMP_ITERATION_TIME_LIMIT_SECS
+                                    && !parts_to_dump.is_empty()
+                                {
+                                    let _timer = metrics::STATE_SYNC_DUMP_ITERATION_ELAPSED
+                                        .with_label_values(&[&shard_id.to_string()])
+                                        .start_timer();
 
-                                            let (part_id, selected_idx) =
-                                                select_random_part_id_with_index(&parts_to_dump);
+                                    let (part_id, selected_idx) =
+                                        select_random_part_id_with_index(&parts_to_dump);
 
-                                            let state_part = match obtain_and_store_state_part(
-                                                runtime.as_ref(),
-                                                shard_id,
-                                                sync_hash,
-                                                &sync_prev_hash,
-                                                &state_root,
-                                                part_id,
-                                                num_parts,
-                                                &chain,
-                                            ) {
-                                                Ok(state_part) => state_part,
-                                                Err(err) => {
-                                                    tracing::warn!(target: "state_sync_dump", shard_id, epoch_height, part_id, ?err, "Failed to obtain and store part. Will skip this part.");
-                                                    break;
-                                                }
-                                            };
-                                            let location = external_storage_location(
-                                                &chain_id,
-                                                &epoch_id,
-                                                epoch_height,
-                                                shard_id,
-                                                part_id,
-                                                num_parts,
-                                            );
-                                            if let Err(_) = external
-                                                .put_state_part(&state_part, shard_id, &location)
-                                                .await
-                                            {
-                                                // no need to break if there's an error, we should keep dumping other parts.
-                                                // reason is we are dumping random selected parts, so it's fine if we are not able to finish all of them
-                                                continue;
-                                            }
-
-                                            // remove the dumped part from parts_to_dump so that we draw without replacement
-                                            parts_to_dump.swap_remove(selected_idx);
-                                            update_dumped_size_and_cnt_metrics(
-                                                &shard_id,
-                                                epoch_height,
-                                                state_part.len(),
-                                            );
+                                    let state_part = match obtain_and_store_state_part(
+                                        runtime.as_ref(),
+                                        shard_id,
+                                        sync_hash,
+                                        &sync_prev_hash,
+                                        &state_root,
+                                        part_id,
+                                        num_parts,
+                                        &chain,
+                                    ) {
+                                        Ok(state_part) => state_part,
+                                        Err(err) => {
+                                            tracing::warn!(target: "state_sync_dump", shard_id, epoch_height, part_id, ?err, "Failed to obtain and store part. Will skip this part.");
+                                            break;
                                         }
-
-                                        if parts_to_dump.is_empty() {
-                                            Ok(Some(StateSyncDumpProgress::AllDumped { epoch_id }))
-                                        } else {
-                                            Ok(Some(StateSyncDumpProgress::InProgress {
-                                                epoch_id,
-                                                epoch_height,
-                                                sync_hash,
-                                            }))
-                                        }
+                                    };
+                                    let location = external_storage_location(
+                                        &chain_id,
+                                        &epoch_id,
+                                        epoch_height,
+                                        shard_id,
+                                        part_id,
+                                        num_parts,
+                                    );
+                                    if let Err(_) = external
+                                        .put_state_part(&state_part, shard_id, &location)
+                                        .await
+                                    {
+                                        // no need to break if there's an error, we should keep dumping other parts.
+                                        // reason is we are dumping random selected parts, so it's fine if we are not able to finish all of them
+                                        continue;
                                     }
+
+                                    // remove the dumped part from parts_to_dump so that we draw without replacement
+                                    parts_to_dump.swap_remove(selected_idx);
+                                    update_dumped_size_and_cnt_metrics(
+                                        &shard_id,
+                                        epoch_height,
+                                        state_part.len(),
+                                    );
+                                }
+                                if parts_to_dump.is_empty() {
+                                    Ok(StateSyncDumpProgress::AllDumped { epoch_id, epoch_height })
+                                } else {
+                                    Ok(StateSyncDumpProgress::InProgress {
+                                        epoch_id,
+                                        epoch_height,
+                                        sync_hash,
+                                    })
                                 }
                             }
                         }
@@ -348,23 +313,11 @@ async fn state_sync_dump(
             }
         };
 
-        // Record the next state of the state machine.
+        // report progress
         let has_progress = match next_state {
-            Ok(Some(next_state)) => {
+            Ok(next_state) => {
                 tracing::debug!(target: "state_sync_dump", shard_id, ?next_state);
-                match chain.store().set_state_sync_dump_progress(shard_id, Some(next_state)) {
-                    Ok(_) => true,
-                    Err(err) => {
-                        // This will be retried.
-                        tracing::debug!(target: "state_sync_dump", shard_id, ?err, "Failed to set progress");
-                        false
-                    }
-                }
-            }
-            Ok(None) => {
-                // Will retry.
-                tracing::debug!(target: "state_sync_dump", shard_id, "Idle");
-                false
+                true
             }
             Err(err) => {
                 // Will retry.
@@ -409,17 +362,7 @@ fn update_dumped_size_and_cnt_metrics(
     metrics::STATE_SYNC_DUMP_NUM_PARTS_DUMPED.with_label_values(&[&shard_id.to_string()]).inc();
 }
 
-fn set_metrics(
-    shard_id: &ShardId,
-    parts_dumped: Option<u64>,
-    num_parts: Option<u64>,
-    epoch_height: Option<EpochHeight>,
-) {
-    if let Some(parts_dumped) = parts_dumped {
-        metrics::STATE_SYNC_DUMP_NUM_PARTS_DUMPED
-            .with_label_values(&[&shard_id.to_string()])
-            .set(parts_dumped as i64);
-    }
+fn set_metrics(shard_id: &ShardId, num_parts: Option<u64>, epoch_height: Option<EpochHeight>) {
     if let Some(num_parts) = num_parts {
         metrics::STATE_SYNC_DUMP_NUM_PARTS_TOTAL
             .with_label_values(&[&shard_id.to_string()])
@@ -428,9 +371,8 @@ fn set_metrics(
     if let Some(epoch_height) = epoch_height {
         assert!(
             epoch_height < 10000,
-            "Impossible: {:?} {:?} {:?} {:?}",
+            "Impossible: {:?} {:?} {:?}",
             shard_id,
-            parts_dumped,
             num_parts,
             epoch_height
         );
@@ -474,7 +416,7 @@ fn start_dumping(
     epoch_manager: &dyn EpochManagerAdapter,
     shard_tracker: &ShardTracker,
     account_id: &Option<AccountId>,
-) -> Result<Option<StateSyncDumpProgress>, Error> {
+) -> Result<StateSyncDumpProgress, Error> {
     let epoch_info = epoch_manager.get_epoch_info(&epoch_id)?;
     let epoch_height = epoch_info.epoch_height();
 
@@ -491,145 +433,41 @@ fn start_dumping(
     let state_header = chain.get_state_response_header(shard_id, sync_hash)?;
     let num_parts = get_num_state_parts(state_header.state_root_node().memory_usage);
     if shard_tracker.care_about_shard(account_id.as_ref(), sync_prev_prev_hash, shard_id, true) {
-        tracing::info!(target: "state_sync_dump", shard_id, ?epoch_id, epoch_height, %sync_prev_hash, %sync_hash, "Initialize dumping state of Epoch");
+        tracing::info!(target: "state_sync_dump", shard_id, ?epoch_id, epoch_height, %sync_prev_hash, %sync_hash, "Dumping state of Epoch");
         // Note that first the state of the state machines gets changes to
         // `InProgress` and it starts dumping state after a short interval.
-        set_metrics(&shard_id, Some(0), Some(num_parts), Some(epoch_height));
-        Ok(Some(StateSyncDumpProgress::InProgress { epoch_id, epoch_height, sync_hash }))
+        set_metrics(&shard_id, Some(num_parts), Some(epoch_height));
+        Ok(StateSyncDumpProgress::InProgress { epoch_id, epoch_height, sync_hash })
     } else {
         tracing::info!(target: "state_sync_dump", shard_id, ?epoch_id, epoch_height, %sync_hash, "Shard is not tracked, skip the epoch");
-        Ok(Some(StateSyncDumpProgress::AllDumped { epoch_id }))
+        Ok(StateSyncDumpProgress::AllDumped { epoch_id, epoch_height })
     }
 }
 
 /// Checks what is the latest complete epoch.
-/// `epoch_id` represents the last fully dumped epoch.
-fn check_new_epoch(
-    epoch_id: Option<EpochId>,
+/// If the shard is tracked, set progress to in progress; else, set progress to fully dumped
+/// return the progress of the dump progress for the shard of the latest complete epoch
+fn check_latest_epoch(
     shard_id: ShardId,
     chain: &Chain,
     epoch_manager: &dyn EpochManagerAdapter,
     shard_tracker: &ShardTracker,
     account_id: &Option<AccountId>,
-) -> Result<Option<StateSyncDumpProgress>, Error> {
+) -> Result<StateSyncDumpProgress, Error> {
     let head = chain.head()?;
-    if Some(&head.epoch_id) == epoch_id.as_ref() {
-        tracing::debug!(target: "state_sync_dump", shard_id, ?epoch_id, "Latest epoch is the same epoch.");
-        Ok(None)
-    } else {
-        // Check if the final block is now in the next epoch.
-        tracing::debug!(target: "state_sync_dump", shard_id, ?epoch_id, "Check if a new complete epoch is available");
-        let hash = head.last_block_hash;
-        let header = chain.get_block_header(&hash)?;
-        let final_hash = header.last_final_block();
-        let sync_hash = StateSync::get_epoch_start_sync_hash(chain, final_hash)?;
-        let header = chain.get_block_header(&sync_hash)?;
-        if Some(header.epoch_id()) == epoch_id.as_ref() {
-            // Still in the latest dumped epoch. Do nothing.
-            Ok(None)
-        } else {
-            start_dumping(
-                head.epoch_id,
-                sync_hash,
-                shard_id,
-                chain,
-                epoch_manager,
-                shard_tracker,
-                account_id,
-            )
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::state_sync::spawn_state_sync_dump;
-    use near_chain::{ChainGenesis, Provenance};
-    use near_chain_configs::{DumpConfig, ExternalStorageLocation};
-    use near_client::sync::state::external_storage_location;
-    use near_client::test_utils::TestEnv;
-    use near_network::test_utils::wait_or_timeout;
-    use near_o11y::testonly::init_test_logger;
-    use near_primitives::types::BlockHeight;
-    use std::ops::ControlFlow;
-    use std::time::Duration;
-
-    #[test]
-    /// Produce several blocks, wait for the state dump thread to notice and
-    /// write files to a temp dir.
-    fn test_state_dump() {
-        init_test_logger();
-
-        let mut chain_genesis = ChainGenesis::test();
-        chain_genesis.epoch_length = 5;
-        let mut env = TestEnv::builder(chain_genesis.clone()).build();
-        let chain = &env.clients[0].chain;
-        let epoch_manager = chain.epoch_manager.clone();
-        let shard_tracker = chain.shard_tracker.clone();
-        let runtime = chain.runtime_adapter.clone();
-        let mut config = env.clients[0].config.clone();
-        let root_dir = tempfile::Builder::new().prefix("state_dump").tempdir().unwrap();
-        config.state_sync.dump = Some(DumpConfig {
-            location: ExternalStorageLocation::Filesystem {
-                root_dir: root_dir.path().to_path_buf(),
-            },
-            restart_dump_for_shards: None,
-            iteration_delay: Some(Duration::from_millis(250)),
-        });
-
-        const MAX_HEIGHT: BlockHeight = 15;
-
-        near_actix_test_utils::run_actix(async move {
-            let _state_sync_dump_handle = spawn_state_sync_dump(
-                &config,
-                chain_genesis,
-                epoch_manager.clone(),
-                shard_tracker.clone(),
-                runtime.clone(),
-                Some("test0".parse().unwrap()),
-            )
-            .unwrap();
-            for i in 1..=MAX_HEIGHT {
-                let block = env.clients[0].produce_block(i as u64).unwrap().unwrap();
-                env.process_block(0, block, Provenance::PRODUCED);
-            }
-            let head = &env.clients[0].chain.head().unwrap();
-            let epoch_id = head.clone().epoch_id;
-            let epoch_info = epoch_manager.get_epoch_info(&epoch_id).unwrap();
-            let epoch_height = epoch_info.epoch_height();
-
-            wait_or_timeout(100, 10000, || async {
-                let mut all_parts_present = true;
-
-                let num_shards = epoch_manager.num_shards(&epoch_id).unwrap();
-                assert_ne!(num_shards, 0);
-
-                for shard_id in 0..num_shards {
-                    let num_parts = 3;
-                    for part_id in 0..num_parts {
-                        let path = root_dir.path().join(external_storage_location(
-                            "unittest",
-                            &epoch_id,
-                            epoch_height,
-                            shard_id,
-                            part_id,
-                            num_parts,
-                        ));
-                        if std::fs::read(&path).is_err() {
-                            println!("Missing {:?}", path);
-                            all_parts_present = false;
-                        }
-                    }
-                }
-                if all_parts_present {
-                    ControlFlow::Break(())
-                } else {
-                    ControlFlow::Continue(())
-                }
-            })
-            .await
-            .unwrap();
-            actix_rt::System::current().stop();
-        });
-    }
+    // Check if the final block is now in the next epoch.
+    tracing::debug!(target: "state_sync_dump", shard_id, "Check if a new complete epoch is available");
+    let hash = head.last_block_hash;
+    let header = chain.get_block_header(&hash)?;
+    let final_hash = header.last_final_block();
+    let sync_hash = StateSync::get_epoch_start_sync_hash(chain, final_hash)?;
+    start_dumping(
+        head.epoch_id,
+        sync_hash,
+        shard_id,
+        chain,
+        epoch_manager,
+        shard_tracker,
+        account_id,
+    )
 }
